2025 09 17

+++ starter/src/compute/compute_init.sh
@@ -1,9 +1,11 @@
-#!/bin/bash
-# compute_java_bootstrap 
+#!/usr/bin/env bash
+# compute_init.sh 
 #
-# Script that is runned once during the setup of a 
-# - compute
-# - with Java
+# Init of a compute
+#
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+cd $SCRIPT_DIR
+
 if [[ -z "$TF_VAR_language" ]]; then
   echo "Missing env variables"
   exit
@@ -40,6 +42,10 @@ install_java() {
       sudo dnf install -y graalvm-21-jdk
       sudo update-alternatives --set java /usr/lib64/graalvm/graalvm-java21/bin/java
       # sudo update-alternatives --set native-image /usr/lib64/graalvm/graalvm-java21/lib/svm/bin/native-image
+    else
+      sudo dnf install -y graalvm-25-jdk
+      sudo update-alternatives --set java /usr/lib64/graalvm/graalvm-java25/bin/java
+      # sudo update-alternatives --set native-image /usr/lib64/graalvm/graalvm-java21/lib/svm/bin/native-image
     fi   
   else
     # JDK 
@@ -51,8 +57,10 @@ install_java() {
       sudo dnf install -y java-11  
     elif [ "$TF_VAR_java_version" == 17 ]; then
       sudo dnf install -y java-17        
+    elif [ "$TF_VAR_java_version" == 21 ]; then
+      sudo dnf install -y java-21         
     else
-      sudo dnf install -y java-21  
+      sudo dnf install -y java-25  
       # Trick to find the path
       # cd -P "/usr/java/latest"
       # export JAVA_LATEST_PATH=`pwd`
@@ -72,66 +80,84 @@ export -f install_java
 # -- App --------------------------------------------------------------------
 # Application Specific installation
 # Build all app* directories
+cd $HOME
+
 for APP_DIR in `ls -d app* | sort -g`; do
   if [ -f $APP_DIR/install.sh ]; then
+    echo "-- $APP_DIR: Install -----------------------------------------"
     chmod +x ${APP_DIR}/install.sh
     ${APP_DIR}/install.sh
   fi  
 done
 
-# -- app/start.sh -----------------------------------------------------------
+# -- app/start*.sh -----------------------------------------------------------
 for APP_DIR in `ls -d app* | sort -g`; do
-  if [ -f $APP_DIR/start.sh ]; then
+  echo "#!/usr/bin/env bash" > $APP_DIR/restart.sh 
+  chmod +x $APP_DIR/restart.sh  
+  for START_SH in `ls $APP_DIR/start*.sh | sort -g`; do
+    echo "-- $START_SH ---------------------------------------"
+    if [[ "$START_SH" =~ start_(.*).sh ]]; then
+      APP_NAME=$(echo "$START_SH" | sed -E 's/(.*)\/start_([a-zA-Z0-9_]+)\.sh$/\1_\2/')
+    else
+      APP_NAME=${APP_DIR}
+    fi
+    echo "APP_NAME=$APP_NAME"
     # Hardcode the connection to the DB in the start.sh
     if [ "$DB_URL" != "" ]; then
-      sed -i "s!##JDBC_URL##!$JDBC_URL!" $APP_DIR/start.sh 
-      sed -i "s!##DB_URL##!$DB_URL!" $APP_DIR/start.sh 
+      sed -i "s!##JDBC_URL##!$JDBC_URL!" $START_SH 
+      sed -i "s!##DB_URL##!$DB_URL!" $START_SH 
     fi  
-    sed -i "s!##TF_VAR_java_vm##!$TF_VAR_java_vm!" $APP_DIR/start.sh   
-    chmod +x $APP_DIR/start.sh
+    sed -i "s!##TF_VAR_java_vm##!$TF_VAR_java_vm!" $START_SH
+    chmod +x $START_SH
 
     # Create an "app.service" that starts when the machine starts.
-    cat > /tmp/$APP_DIR.service << EOT
+     cat > /tmp/$APP_NAME.service << EOT
 [Unit]
 Description=App
 After=network.target
 
 [Service]
 Type=simple
-ExecStart=/home/opc/$APP_DIR/start.sh
+ExecStart=/home/opc/$START_SH
 TimeoutStartSec=0
 User=opc
 
 [Install]
 WantedBy=default.target
 EOT
-
-    sudo cp /tmp/$APP_DIR.service /etc/systemd/system
-    sudo chmod 664 /etc/systemd/system/$APP_DIR.service
+    sudo cp /tmp/$APP_NAME.service /etc/systemd/system
+    sudo chmod 664 /etc/systemd/system/$APP_NAME.service
     sudo systemctl daemon-reload
-    sudo systemctl enable $APP_DIR.service
-    sudo systemctl restart $APP_DIR.service
-  fi
-done  
+    sudo systemctl enable $APP_NAME.service
+    sudo systemctl restart $APP_NAME.service
+    echo "sudo systemctl restart $APP_NAME" >> $APP_DIR/restart.sh 
+  done  
+done 
+
+# -- Helper --------------------------------------------------------------------
+cd $SCRIPT_DIR
+mv helper.sh $HOME
 
 # -- UI --------------------------------------------------------------------
 # Install NGINX
 sudo dnf install nginx -y > /tmp/dnf_nginx.log
 
 # Default: location /app/ { proxy_pass http://localhost:8080 }
-sudo cp nginx_app.locations /etc/nginx/conf.d/.
-if grep -q nginx_app /etc/nginx/nginx.conf; then
-  echo "Include nginx_app.locations is already there"
-else
+if [ -f nginx_app.locations ]; then
+  sudo cp nginx_app.locations /etc/nginx/conf.d/.
+  if grep -q nginx_app /etc/nginx/nginx.conf; then
+    echo "Include nginx_app.locations is already there"
+  else
     echo "Adding nginx_app.locations"
     sudo awk -i inplace '/404.html/ && !x {print "        include conf.d/nginx_app.locations;"; x=1} 1' /etc/nginx/nginx.conf
+  fi
 fi
 
 # TLS
 if [ -f nginx_tls.conf ]; then
-    echo "Adding nginx_tls.conf"
-    sudo cp nginx_tls.conf /etc/nginx/conf.d/.
-    sudo awk -i inplace '/# HTTPS server/ && !x {print "        include conf.d/nginx_tls.conf;"; x=1} 1' /etc/nginx/nginx.conf
+  echo "Adding nginx_tls.conf"
+  sudo cp nginx_tls.conf /etc/nginx/conf.d/.
+  sudo awk -i inplace '/# HTTPS server/ && !x {print "        include conf.d/nginx_tls.conf;"; x=1} 1' /etc/nginx/nginx.conf
 fi
 
 # SE Linux (for proxy_pass)
@@ -141,6 +167,7 @@ sudo setsebool -P httpd_can_network_connect 1
 sudo systemctl enable nginx
 sudo systemctl restart nginx
 
+cd $HOME
 if [ -d ui ]; then
   # Copy the index file after the installation of nginx
   sudo cp -r ui/* /usr/share/nginx/html/
